SYNC-1750 Avoid code duplication
