+++ modules/apps/sync/sync-service/src/main/java/com/liferay/sync/internal/model/listener/SyncBaseModelListener.java
@@ -16,6 +16,13 @@ package com.liferay.sync.internal.model.listener;
 
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.document.library.kernel.model.DLFolder;
+import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Disjunction;
+import com.liferay.portal.kernel.dao.orm.DynamicQuery;
+import com.liferay.portal.kernel.dao.orm.Property;
+import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
+import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
+import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.BaseModel;
 import com.liferay.portal.kernel.model.BaseModelListener;
 import com.liferay.portal.kernel.model.ResourcePermission;
@@ -25,6 +32,7 @@ import com.liferay.sync.constants.SyncDLObjectConstants;
 import com.liferay.sync.model.SyncDLObject;
 import com.liferay.sync.service.SyncDLObjectLocalService;
 
+import java.util.Date;
 import java.util.List;
 
 import org.osgi.service.component.annotations.Reference;
@@ -54,6 +62,77 @@ public abstract class SyncBaseModelListener<T extends BaseModel<T>>
 		return null;
 	}
 
+	protected void onAddRoleAssociation(Object roleId)
+		throws ModelListenerException {
+
+		ActionableDynamicQuery actionableDynamicQuery =
+			_getActionableDynamicQuery(roleId);
+
+		actionableDynamicQuery.setPerformActionMethod(
+			new ActionableDynamicQuery.
+				PerformActionMethod<ResourcePermission>() {
+
+				@Override
+				public void performAction(
+					ResourcePermission resourcePermission) {
+
+					SyncDLObject syncDLObject = getSyncDLObject(
+						resourcePermission);
+
+					if (syncDLObject == null) {
+						return;
+					}
+
+					updateSyncDLObject(syncDLObject);
+				}
+
+			});
+
+		try {
+			actionableDynamicQuery.performActions();
+		}
+		catch (Exception e) {
+			throw new ModelListenerException(e);
+		}
+	}
+
+	protected void onRemoveRoleAssociation(Object roleId) {
+		ActionableDynamicQuery actionableDynamicQuery =
+			_getActionableDynamicQuery(roleId);
+
+		actionableDynamicQuery.setPerformActionMethod(
+			new ActionableDynamicQuery.
+				PerformActionMethod<ResourcePermission>() {
+
+				@Override
+				public void performAction(
+					ResourcePermission resourcePermission) {
+
+					SyncDLObject syncDLObject = getSyncDLObject(
+						resourcePermission);
+
+					if (syncDLObject == null) {
+						return;
+					}
+
+					Date date = new Date();
+
+					syncDLObject.setModifiedTime(date.getTime());
+					syncDLObject.setLastPermissionChangeDate(date);
+
+					syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
+				}
+
+			});
+
+		try {
+			actionableDynamicQuery.performActions();
+		}
+		catch (Exception e) {
+			throw new ModelListenerException(e);
+		}
+	}
+
 	protected void updateSyncDLObject(SyncDLObject syncDLObject) {
 		syncDLObject.setModifiedTime(System.currentTimeMillis());
 
@@ -79,5 +158,40 @@ public abstract class SyncBaseModelListener<T extends BaseModel<T>>
 
 	@Reference
 	protected SyncDLObjectLocalService syncDLObjectLocalService;
+
+	private ActionableDynamicQuery _getActionableDynamicQuery(
+		final Object roleId) {
+
+		ActionableDynamicQuery actionableDynamicQuery =
+			resourcePermissionLocalService.getActionableDynamicQuery();
+
+		actionableDynamicQuery.setAddCriteriaMethod(
+			new ActionableDynamicQuery.AddCriteriaMethod() {
+
+				@Override
+				public void addCriteria(DynamicQuery dynamicQuery) {
+					Property nameProperty = PropertyFactoryUtil.forName("name");
+					Property roleIdProperty = PropertyFactoryUtil.forName(
+						"roleId");
+					Property viewActionIdProperty = PropertyFactoryUtil.forName(
+						"viewActionId");
+
+					Disjunction disjunction =
+						RestrictionsFactoryUtil.disjunction();
+
+					disjunction.add(
+						nameProperty.eq(DLFileEntry.class.getName()));
+					disjunction.add(nameProperty.eq(DLFolder.class.getName()));
+
+					dynamicQuery.add(disjunction);
+
+					dynamicQuery.add(roleIdProperty.eq(roleId));
+					dynamicQuery.add(viewActionIdProperty.eq(true));
+				}
+
+			});
+
+		return actionableDynamicQuery;
+	}
 
 }
