+++ modules/apps/sync/sync-service/src/main/java/com/liferay/sync/internal/model/listener/UserModelListener.java
@@ -14,27 +14,16 @@
 
 package com.liferay.sync.internal.model.listener;
 
-import com.liferay.document.library.kernel.model.DLFileEntry;
-import com.liferay.document.library.kernel.model.DLFolder;
-import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
-import com.liferay.portal.kernel.dao.orm.Disjunction;
-import com.liferay.portal.kernel.dao.orm.DynamicQuery;
-import com.liferay.portal.kernel.dao.orm.Property;
-import com.liferay.portal.kernel.dao.orm.PropertyFactoryUtil;
 import com.liferay.portal.kernel.dao.orm.QueryUtil;
-import com.liferay.portal.kernel.dao.orm.RestrictionsFactoryUtil;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.model.ModelListener;
-import com.liferay.portal.kernel.model.ResourcePermission;
 import com.liferay.portal.kernel.model.Role;
 import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.sync.constants.SyncDeviceConstants;
-import com.liferay.sync.model.SyncDLObject;
 import com.liferay.sync.model.SyncDevice;
 import com.liferay.sync.service.SyncDeviceLocalService;
 
-import java.util.Date;
 import java.util.List;
 
 import org.osgi.service.component.annotations.Component;
@@ -69,63 +58,8 @@ public class UserModelListener extends SyncBaseModelListener<User> {
 			Object associationClassPK)
 		throws ModelListenerException {
 
-		if (!associationClassName.equals(Role.class.getName())) {
-			return;
-		}
-
-		ActionableDynamicQuery actionableDynamicQuery =
-			resourcePermissionLocalService.getActionableDynamicQuery();
-
-		actionableDynamicQuery.setAddCriteriaMethod(
-			new ActionableDynamicQuery.AddCriteriaMethod() {
-
-				@Override
-				public void addCriteria(DynamicQuery dynamicQuery) {
-					Property nameProperty = PropertyFactoryUtil.forName("name");
-					Property roleIdProperty = PropertyFactoryUtil.forName(
-						"roleId");
-					Property viewActionIdProperty = PropertyFactoryUtil.forName(
-						"viewActionId");
-
-					Disjunction disjunction =
-						RestrictionsFactoryUtil.disjunction();
-
-					disjunction.add(
-						nameProperty.eq(DLFileEntry.class.getName()));
-					disjunction.add(nameProperty.eq(DLFolder.class.getName()));
-
-					dynamicQuery.add(disjunction);
-
-					dynamicQuery.add(roleIdProperty.eq(associationClassPK));
-					dynamicQuery.add(viewActionIdProperty.eq(true));
-				}
-
-			});
-		actionableDynamicQuery.setPerformActionMethod(
-			new ActionableDynamicQuery.
-				PerformActionMethod<ResourcePermission>() {
-
-				@Override
-				public void performAction(
-					ResourcePermission resourcePermission) {
-
-					SyncDLObject syncDLObject = getSyncDLObject(
-						resourcePermission);
-
-					if (syncDLObject == null) {
-						return;
-					}
-
-					updateSyncDLObject(syncDLObject);
-				}
-
-			});
-
-		try {
-			actionableDynamicQuery.performActions();
-		}
-		catch (Exception e) {
-			throw new ModelListenerException(e);
+		if (associationClassName.equals(Role.class.getName())) {
+			onAddRoleAssociation(associationClassPK);
 		}
 	}
 
@@ -135,68 +69,8 @@ public class UserModelListener extends SyncBaseModelListener<User> {
 			Object associationClassPK)
 		throws ModelListenerException {
 
-		if (!associationClassName.equals(Role.class.getName())) {
-			return;
-		}
-
-		ActionableDynamicQuery actionableDynamicQuery =
-			resourcePermissionLocalService.getActionableDynamicQuery();
-
-		actionableDynamicQuery.setAddCriteriaMethod(
-			new ActionableDynamicQuery.AddCriteriaMethod() {
-
-				@Override
-				public void addCriteria(DynamicQuery dynamicQuery) {
-					Property nameProperty = PropertyFactoryUtil.forName("name");
-					Property roleIdProperty = PropertyFactoryUtil.forName(
-						"roleId");
-					Property viewActionIdProperty = PropertyFactoryUtil.forName(
-						"viewActionId");
-
-					Disjunction disjunction =
-						RestrictionsFactoryUtil.disjunction();
-
-					disjunction.add(
-						nameProperty.eq(DLFileEntry.class.getName()));
-					disjunction.add(nameProperty.eq(DLFolder.class.getName()));
-
-					dynamicQuery.add(disjunction);
-
-					dynamicQuery.add(roleIdProperty.eq(associationClassPK));
-					dynamicQuery.add(viewActionIdProperty.eq(true));
-				}
-
-			});
-		actionableDynamicQuery.setPerformActionMethod(
-			new ActionableDynamicQuery.
-				PerformActionMethod<ResourcePermission>() {
-
-				@Override
-				public void performAction(
-					ResourcePermission resourcePermission) {
-
-					SyncDLObject syncDLObject = getSyncDLObject(
-						resourcePermission);
-
-					if (syncDLObject == null) {
-						return;
-					}
-
-					Date date = new Date();
-
-					syncDLObject.setModifiedTime(date.getTime());
-					syncDLObject.setLastPermissionChangeDate(date);
-
-					syncDLObjectLocalService.updateSyncDLObject(syncDLObject);
-				}
-
-			});
-
-		try {
-			actionableDynamicQuery.performActions();
-		}
-		catch (Exception e) {
-			throw new ModelListenerException(e);
+		if (associationClassName.equals(Role.class.getName())) {
+			onRemoveRoleAssociation(associationClassPK);
 		}
 	}
